Subject: Updates.
Description:

1. Population of users and codes separated.
2. FIX: alignment of Github-ribbon when no code is submitted.
3. Alignment in registration template.

+++ populate_haqiba.py
@@ -3,10 +3,14 @@ import os
 import sys
 import store
 
-def populate():
+def populate_users():
 
+    # Admin 
+    os.system("python manage.py syncdb --noinput")
+    os.system("python manage.py createsuperuser --username=admin --email=admin@example.com")
+
+    # Normal users
     store.USERNAME = add_user(store.USERNAME, store.EMAIL, store.PASSWORD)
-
     
     add_user_profile(user=store.USERNAME, 
                      website=store.WEBSITE,
@@ -21,6 +25,14 @@ def populate():
                          picture=store.PHOTO_2)
     except:
         print "No USERNAME_2."
+
+    user_list = User.objects.all()
+    if user_list:
+        print "Following user(s) created successfully."
+        for i in user_list:
+            print i.username
+
+def populate_codes():
     
     add_code_template(
         user_id=store.USERNAME.id,
@@ -54,7 +66,7 @@ buffer is not visiting a file."
                          (ido-read-file-name "Find file(as root): ")))
     (find-alternate-file (concat "/sudo:root@localhost:" buffer-file-name))))
 (global-set-key (kbd "C-x C-r") 'sudo-edit)""",
-        description="Edit currently visited file as root. Keybinding: C-x C-r",
+        description="Edit currently visited file as root. Key binding: C-x C-r",
         screenshot='screenshot/banner.png'
     )
 
@@ -157,7 +169,7 @@ buffer is not visiting a file."
 				(emacs-lisp . t)
 				(clojure . t)
 				(C . t)))""",
-        description="Provide language specific ayntax highlighting when converting org to PDF.",
+        description="Provide language specific syntax highlighting when converting org to PDF.",
         screenshot='screenshot/banner.png'
     )
 
@@ -258,7 +270,7 @@ buffer is not visiting a file."
 		 (pluralize "word" count))))))
 (global-set-key (kbd "C-x c") 'count-words-buffer)
 """,
-        description="Count total number of words in current buffer. Keybinding: C-x c.",
+        description="Count total number of words in current buffer. Key binding: C-x c.",
         screenshot="/screenshot/banner.png"
     )
 
@@ -278,7 +290,7 @@ buffer is not visiting a file."
         (ansi-term (getenv "SHELL")))
     (switch-to-buffer-other-window "*ansi-term*")))
 (global-set-key (kbd "C-c t") 'visit-term-buffer)""",
-        description="Visit terminal buffer. Keybinding: C-c t. From http://emacsredux.com/blog/page/2/",
+        description="Visit terminal buffer. Key binding: C-c t. From http://emacsredux.com/blog/page/2/",
         screenshot="/screenshot/banner.png"
     )    
 
@@ -291,7 +303,7 @@ buffer is not visiting a file."
   (interactive "nPercent: ")
   (goto-char (/ (* (point-max) pct) 100)))
 (global-set-key (kbd "C-x p") 'goto-percent)""",
-        description="Go to place in a buffer expressed in percentage. Keybinding: C-x p.",
+        description="Go to place in a buffer expressed in percentage. Key binding: C-x p.",
         screenshot="/screenshot/banner.png"
     )
 
@@ -375,6 +387,7 @@ programming."
     except:
         print "No USERNAME_2 available for this code."
 
+    print "---------- Codes ------------"
     for c in CodeTemplate.objects.all():
         print c.name
 
@@ -400,7 +413,9 @@ if __name__ == '__main__':
     os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'haqiba.settings')
     from emacshaqiba.models import CodeTemplate, UserProfile
     from django.contrib.auth.models import User
-    os.system("rm haqiba.db")
-    os.system("python manage.py syncdb --noinput")
-    os.system("python manage.py createsuperuser --username=admin --email=admin@example.com")
-    populate()
+
+    if os.path.exists('haqiba.db'):
+        os.system("rm haqiba.db")
+
+    populate_users()
+    populate_codes()
